+++ NEWS.md
@@ -1,5 +1,32 @@
 # News 
 
+## Development
+
+New features:
+
+* `modelsummary()` gets a `gof_function` argument which accepts functions to extract custom information from models.
+* `flextable`: Support spanning column headers
+
+`datasummary_balance()`:
+
+* `formula` can now include variables on the left-hand side to indicate the subset of columns to summarize: `datasummary_balance(mpg + hp ~ am, data = mtcars)` Thanks to @etiennebacher for feature request #751.
+
+Minor:
+
+* Unnecessary text printed to screen on some F stat computations is now suppressed.
+* Update to `tinytable` 0.3.0
+
+Bugs:
+
+* `escape` argument not respected in `datasummary_df()`. Thanks to @adamaltmejd for report #740
+* `datasummary_correlation()` supports `data.table`. Thanks to volatilehead on Twitter for report #737.
+* Accepts named `estimate` argument when using shape and statistics are horizontal. Thanks to @iago-pssjd for report #745.
+* Labelled values but no label for variable broke `datasummary()`. Thanks to @marklhc for report #752.
+* `coef_map` does not work when there is a `group`. Thanks to @mccarthy-m-g for report #757.
+* `kableExtra`: fix spanning column headers when using the `shape` argument.
+* Multiple footnotes and line breaks in footnotes are now allowed in `tinytable` output. Thanks to 
+
+
 ## 2.0.0
 
 MAJOR BREAKING CHANGE: The default output format is now `tinytable` instead of `kableExtra`. Learn more about `tinytable` here:
@@ -24,16 +51,20 @@ New features:
 
 * `datasummary_skim()`:
   - New `type="all"` by default to display both numeric and categorical variables in a single table with distinct panels. This feature is only available with the `tinytable` backend (default).
+  - `by` argument allows summarizing numeric variables by group.
   - `fun_numeric` argument accepts a list of functions to control the summary columns.
 * `modelsummary()`:
   - `statistic` and `estimate` can be specified as named vectors to control the names of statistics when displayed in different columns using the `shape` argument. (Thanks to @mps9506 for bug report #722)
   - `modelsummary(panels, shape = "cbind")` automatically adds column-spanning labels when `panels` is a named nested list of models.
 * `config_modelsummary()` gets a `startup_message` argument to silence the startup message persistently.
+* Improved documentation and vignettes, providing clearer instructions and examples.
+* Updated tests and snapshots to ensure reliability and consistency across changes.
 
 Bug fixes:
 
 * Fixed Issue #399: datasummary_balance() siunitx formatting.
 * Fixed Issue #782: Useless warning in some `modelplot()` calls. Thanks to @iago-pssjd for the report and @florence-laflamme for the fix.
+* Addressed various bugs and made optimizations for better performance and user experience.
 
 
 ## 1.4.5